fudbal/requirements.txt
aiohttp==3.10.9
colorama==0.4.6
httpx==0.27.2
pandas==2.2.2
python-telegram-bot==21.6
pytz==2024.2
RapidFuzz==3.10.0
requests==2.32.3
pandas==2.2.2
